requirements.txt
fastapi==0.116.1
fastapi_jwt_auth==0.5.0
httpx==0.28.1
pydantic==1.10.22
pymongo==4.13.2
pytest==8.4.1
python-dotenv==1.1.1
